Visualization page re-factoring; Add kills-by-weapon bar chart
