+++ .babelrc
@@ -1,4 +1,4 @@
 {
-  presets: ['@babel/preset-react', '@babel/preset-env'],
+  presets: ['@babel/preset-react', '@babel/preset-env', '@babel/preset-flow'],
   plugins: ['@babel/plugin-proposal-class-properties', 'babel-plugin-styled-components']
 }
