+++ .gitignore
@@ -1,3 +1,5 @@
 node_modules/
 .tern-port
 dist/
+.vscode
+.idea
