+++ package.json
@@ -37,6 +37,7 @@
     "@babel/plugin-proposal-class-properties": "^7.1.0",
     "@babel/polyfill": "^7.0.0",
     "@babel/preset-env": "^7.1.6",
+    "@babel/preset-flow": "^7.0.0",
     "@babel/preset-react": "^7.0.0",
     "@babel/register": "^7.0.0",
     "babel-eslint": "^10.0.1",
@@ -44,8 +45,12 @@
     "babel-plugin-styled-components": "^1.8.0",
     "css-loader": "^1.0.1",
     "eslint": "^5.9.0",
+    "eslint-config-prettier": "^4.1.0",
+    "eslint-plugin-flowtype": "^3.6.0",
     "eslint-plugin-react": "^7.11.1",
     "file-loader": "^2.0.0",
+    "flow-bin": "^0.96.0",
+    "flow-typed": "^2.5.1",
     "html-loader": "^0.5.5",
     "markdown-loader": "^4.0.0",
     "marked": "^0.5.1",
