+++ src/front/src/pages/visualization/VizTypeSelect.js
@@ -1,0 +1,15 @@
+import React from 'react';
+
+import { Select } from 'common/ui';
+
+const VizTypeSelect = ({ handleChange, value, fluid, options }) => (
+  <Select
+    options={options}
+    fluid={fluid}
+    handleChange={handleChange}
+    value={value}
+    aria-label="visualization-type"
+  />
+);
+
+export default VizTypeSelect;
