+++ src/front/src/pages/visualization/charts/heatmap/index.js
@@ -1,0 +1,3 @@
+import HeatmapChart from './HeatmapContainer';
+
+export default HeatmapChart;
