+++ src/front/src/pages/visualization/charts/index.js
@@ -1,0 +1,2 @@
+export { default as HeatmapChart } from './heatmap';
+export { default as BarChart } from './bar';
